requirements.txt
streamlit==1.27.0
pandas==2.1.0
python-dotenv==1.0.0